+++ include/TileMap.hpp
@@ -122,7 +122,7 @@ class Layer
     Layer() = default;
     virtual ~Layer() = default;
 
-    virtual void render() = 0;
+    virtual void draw() = 0;
     virtual void setOpacity(double value) = 0;
     virtual double getOpacity() const = 0;
 
@@ -179,7 +179,7 @@ class TileLayer : public Layer
     [[nodiscard]] std::vector<TileResult> getFromArea(const Rect& area) const;
     [[nodiscard]] std::optional<TileResult> getFromPoint(const Vec2& position) const;
 
-    void render() override;
+    void draw() override;
     void setOpacity(double value) override;
     double getOpacity() const override;
 
@@ -245,7 +245,7 @@ class ObjectGroup : public Layer
 
     [[nodiscard]] tmx::ObjectGroup::DrawOrder getDrawOrder() const;
     [[nodiscard]] const std::vector<MapObject>& getObjects() const;
-    void render() override;
+    void draw() override;
     void setOpacity(double value) override;
     double getOpacity() const override;
 
@@ -265,7 +265,7 @@ class ImageLayer : public Layer
     ~ImageLayer() = default;
 
     [[nodiscard]] std::shared_ptr<Texture> getTexture() const;
-    void render() override;
+    void draw() override;
     void setOpacity(double value) override;
     double getOpacity() const override;
 
@@ -283,7 +283,7 @@ class Map
     Map() = default;
     ~Map() = default;
 
-    void render();
+    void draw();
     void load(const std::string& tmxPath);
 
     [[nodiscard]] tmx::Orientation getOrientation() const;